+++ view/solicitud_tecnico/solicitud.List.php
@@ -1,8 +1,8 @@
 <!--autor: Sellan Fajardo Leonardo-->
-<?php 
-$link ="index.php?c=Login&f=index";
+<?php
+$link = "index.php?c=Login&f=index";
 $imagen = "assets/imagenes/salir.png";
-$opcion ="Salir";
+$opcion = "Salir";
 $titulo = "Lista de Solicitudes tecnicos";
 require_once HEADER; ?>
 
@@ -108,13 +108,17 @@ require_once HEADER; ?>
             resultados += '<td>' + solicitud[i].correo + '</td>';
             resultados += '<td>' + solicitud[i].fecha_solicitud + '</td>';
             resultados += '<td>' + solicitud[i].id_problemas + '</td>';
-            resultados += '<td>' +
-                "<a href='index.php?c=solicitudtecnico&a=editar&id=" + solicitud[i].id_solicitud +
-                "' " + "class='btn btn-primary'><i class='fas fa-marker'></i></a>" +
-                "<a href='index.php?c=solicitudtecnico&a=eliminar&id=" + solicitud[i].id_solicitud + "'" +
-                "class='btn btn-danger' onclick = 'if (!confirm(\'Desea eliminar la actividad: '" + solicitud[i].nombre +
-                " \')) return false; " + " ><i class='far fa-trash-alt'></i> </a>" + '</td>';
-            resultados += '</tr>';
+            <?php if ($_SESSION['rol'] == null || $_SESSION['rol'] == 1 || $_SESSION['rol'] == 3) { ?>
+                resultados += '<td>' +
+                    "<a href='index.php?c=solicitudtecnico&a=editar&id=" + solicitud[i].id_solicitud +
+                    "' " + "class='btn btn-primary'><i class='fas fa-marker'></i></a>" +
+                <?php } ?>
+                <?php if ($_SESSION['rol'] == 3) { ?>
+                        "<a href='index.php?c=solicitudtecnico&a=eliminar&id=" + solicitud[i].id_solicitud + "'" +
+                        "class='btn btn-danger' onclick = 'if (!confirm(\'Desea eliminar la actividad: '" + solicitud[i].nombre +
+                        " \')) return false; " + " ><i class='far fa-trash-alt'></i> </a>" + '</td>';
+                    resultados += '</tr>';
+                <?php } ?>
         }
         tbody.innerHTML = resultados;
     }
